Invalidate queries after a transaction is sent or confirmed
This is necessary to make sure that transaction lists are updated when a new one is sent. By invalidating the queries, the transactions are refetched and replace the ones in the cache.

Add `useAddress` hook and use it instead of the `useSafeInfo` when only the address is needed.

Move the queryClient and invalidateQueries functions to a new file called queryClient.ts for better organization and separation of concerns.